+++ src/index.js
@@ -1,8 +1,11 @@
 const { getValidators, registerValidator } = require('./validators')
+const { registerTranslation, setTranslation } = require('./translation')
 const validateSchema = require('./validate-schema')
 
 module.exports = {
   getValidators,
   registerValidator,
-  validateSchema
+  validateSchema,
+  registerTranslation,
+  setTranslation
 }