+++ tasks/generate-docs/generate-docs.js
@@ -1,0 +1,3 @@
+const generateValidators = require('./validators')
+
+generateValidators()